fix: ensure that existing rule set can be used
The `static` method exposed by `@aws-cdk/aws-ses`, to retrieve an existing rule set, `ReceiptRuleSet.fromReceiptRuleSetName(Construct, string, string)` returns a behavioural interface, as opposed to a concrete implementation. By enforcing that a concrete instance be supplied as a property to the `EmailForwardingRuleSet`, it is not possible to supply an existing rule set created externally from the stack.